compliance_platform/requirements.txt
boto3>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
python-dotenv>=1.0.0